+++ src/fosanalysis/preprocessing/resizing/__init__.py
@@ -1,0 +1,14 @@
+
+"""
+Contains modules for resizing data:
+- downsampling: Combining several readings into one using aggregate functions,
+- resampling: changing spatial or temporal spacing of data using interpolation.
+
+\author Bertram Richter
+\date 2024
+"""
+
+from . import aggregate
+from . import downsampling
+from . import resampling
+